requirements.txt
luqum
elasticsearch-dsl
pyyaml
